split out iam user to it's own function, add the ability to pull user tags
